requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2==2.9.1
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
